telegram-bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.25.2